+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 * Remove builder pattern because `TileJSON` is writable 
 * Add `other` fields for any unknown fields in root and vector layers
 * Restructure instantiation:
-  * use `new(source)` or `new_ext(sources, version)` to create `TileJSON`
+  * use `tilejson!{ source }` macro to create `TileJSON` objects, with any number of the optional `field: value` pairs. 
   * use `set_missing_defaults()` to replace all missing values with their defaults (only if the spec defines it)
 * Remove `id` field because it is not supported by the spec